+++ changelog.md
@@ -2,8 +2,9 @@
 
 ## TBD
 
+- Added Volcanic Ammolet
 - Added masteries for Vacuum Cleaner, Paintball Cannon, Gunbrella, Alyx, and Pistol Whip
-- Added 5 new synergies
+- Added 6 new synergies
 - Added indicator for amount of debris vacuumed to Vacuum Cleaner ammo display
 - Vacuum Cleaner debris is now processed when it is actually absorbed, rather than when it is suctioned
 - Tweaked Gunbrella projectiles so that they actually fall closer to the reticle rather than overshooting it